Moved data crosswalks to inst/extdata
keep cleaned, final criteria data tables in data/ folder

put raw data files in inst/extdata/ (those that need validating)

wrote a test function to validate ATTAINS.ParameterName to see if they are found as an ATTAINS domain value. For this test, we can allow these entries to still be entered, but warn users that these will likely need to be requested as new ATTAINS.ParameterName is one route we can take.